Updated createBody, weather, and getWeather functions in src/body.ts and src/weather.ts
Changes made in createBody function:

Added background styling to the body element.
Created an h1 element with the text "Weather Forecast" and added styling to it.
Created a div element and added styling to it.
Created an input element with search functionality and added styling to it.
Created a button element with a search icon and added styling to it.

Changes made in getWeather function:

Added an async keyword to indicate that the function is asynchronous.
Used the fetch function to make an API call to retrieve weather data based on the provided city.
Retrieved the JSON response data using the response.json() method.
Created a weatherObj object using the weather function.
Assigned the retrieved weather data to the respective properties of the weatherObj object.

These changes were made to improve the functionality and styling of the createBody function, and to properly handle and store weather data in the getWeather function.

[Ticket: X]